+++ include/pisa/bit_vector.hpp
@@ -9,184 +9,7 @@
 
 namespace pisa {
 
-namespace detail {
-    inline size_t words_for(uint64_t n) { return ceil_div(n, 64); }
-}  // namespace detail
-
-class bit_vector_builder {
-  public:
-    using bits_type = std::vector<uint64_t>;
-
-    explicit bit_vector_builder(uint64_t size = 0, bool init = false) : m_size(size)
-    {
-        m_bits.resize(detail::words_for(size), init ? std::numeric_limits<std::uint64_t>::max() : 0U);
-        if (size != 0U) {
-            m_cur_word = &m_bits.back();
-            // clear padding bits
-            if (init && ((size % 64) != 0U)) {
-                *m_cur_word >>= 64 - (size % 64);
-            }
-        }
-    }
-    bit_vector_builder(bit_vector_builder const&) = delete;
-    bit_vector_builder(bit_vector_builder&&) = delete;
-    bit_vector_builder& operator=(bit_vector_builder const&) = delete;
-    bit_vector_builder& operator=(bit_vector_builder&&) = delete;
-    ~bit_vector_builder() = default;
-
-    void reserve(uint64_t size) { m_bits.reserve(detail::words_for(size)); }
-
-    inline void push_back(bool b)
-    {
-        uint64_t pos_in_word = m_size % 64;
-        if (pos_in_word == 0) {
-            m_bits.push_back(0);
-            m_cur_word = &m_bits.back();
-        }
-        *m_cur_word |= (uint64_t)b << pos_in_word;
-        ++m_size;
-    }
-
-    inline void set(uint64_t pos, bool b)
-    {
-        uint64_t word = pos / 64;
-        uint64_t pos_in_word = pos % 64;
-
-        m_bits[word] &= ~(uint64_t(1) << pos_in_word);
-        m_bits[word] |= uint64_t(b) << pos_in_word;
-    }
-
-    inline void set_bits(uint64_t pos, uint64_t bits, size_t len)
-    {
-        assert(pos + len <= size());
-        // check there are no spurious bits
-        assert(len == 64 || (bits >> len) == 0);
-        if (len == 0U) {
-            return;
-        }
-        uint64_t mask = (len == 64) ? uint64_t(-1) : ((uint64_t(1) << len) - 1);
-        uint64_t word = pos / 64;
-        uint64_t pos_in_word = pos % 64;
-
-        m_bits[word] &= ~(mask << pos_in_word);
-        m_bits[word] |= bits << pos_in_word;
-
-        uint64_t stored = 64 - pos_in_word;
-        if (stored < len) {
-            m_bits[word + 1] &= ~(mask >> stored);
-            m_bits[word + 1] |= bits >> stored;
-        }
-    }
-
-    inline void append_bits(uint64_t bits, size_t len)
-    {
-        // check there are no spurious bits
-        assert(len == 64 || (bits >> len) == 0);
-        if (len == 0U) {
-            return;
-        }
-        uint64_t pos_in_word = m_size % 64;
-        m_size += len;
-        if (pos_in_word == 0) {
-            m_bits.push_back(bits);
-        } else {
-            *m_cur_word |= bits << pos_in_word;
-            if (len > 64 - pos_in_word) {
-                m_bits.push_back(bits >> (64 - pos_in_word));
-            }
-        }
-        m_cur_word = &m_bits.back();
-    }
-
-    inline void zero_extend(uint64_t n)
-    {
-        m_size += n;
-        uint64_t needed = detail::words_for(m_size) - m_bits.size();
-        if (needed != 0U) {
-            m_bits.insert(m_bits.end(), needed, 0);
-            m_cur_word = &m_bits.back();
-        }
-    }
-
-    inline void one_extend(uint64_t n)
-    {
-        while (n >= 64) {
-            append_bits(uint64_t(-1), 64);
-            n -= 64;
-        }
-        if (n != 0U) {
-            append_bits(uint64_t(-1) >> (64 - n), n);
-        }
-    }
-
-    void append(bit_vector_builder const& rhs)
-    {
-        if (rhs.size() == 0U) {
-            return;
-        }
-
-        uint64_t pos = m_bits.size();
-        uint64_t shift = size() % 64;
-        m_size = size() + rhs.size();
-        m_bits.resize(detail::words_for(m_size));
-
-        if (shift == 0) {  // word-aligned, easy case
-            std::copy(rhs.m_bits.begin(), rhs.m_bits.end(), m_bits.begin() + ptrdiff_t(pos));
-        } else {
-            uint64_t* cur_word = &m_bits.front() + pos - 1;
-            for (size_t i = 0; i < rhs.m_bits.size() - 1; ++i) {
-                uint64_t w = rhs.m_bits[i];
-                *cur_word |= w << shift;
-                *++cur_word = w >> (64 - shift);
-            }
-            *cur_word |= rhs.m_bits.back() << shift;
-            if (cur_word < &m_bits.back()) {
-                *++cur_word = rhs.m_bits.back() >> (64 - shift);
-            }
-        }
-        m_cur_word = &m_bits.back();
-    }
-
-    // reverse in place
-    void reverse()
-    {
-        uint64_t shift = 64 - (size() % 64);
-
-        uint64_t remainder = 0;
-        for (auto& word: m_bits) {
-            uint64_t cur_word;
-            if (shift != 64) {  // this should be hoisted out
-                cur_word = remainder | (word << shift);
-                remainder = word >> (64 - shift);
-            } else {
-                cur_word = word;
-            }
-            word = broadword::reverse_bits(cur_word);
-        }
-        assert(remainder == 0);
-        std::reverse(m_bits.begin(), m_bits.end());
-    }
-
-    bits_type& move_bits()
-    {
-        assert(detail::words_for(m_size) == m_bits.size());
-        return m_bits;
-    }
-
-    uint64_t size() const { return m_size; }
-
-    void swap(bit_vector_builder& other)
-    {
-        m_bits.swap(other.m_bits);
-        std::swap(m_size, other.m_size);
-        std::swap(m_cur_word, other.m_cur_word);
-    }
-
-  private:
-    bits_type m_bits;
-    uint64_t m_size;
-    uint64_t* m_cur_word;
-};
+class bit_vector_builder;
 
 class bit_vector {
   public:
@@ -218,11 +41,7 @@ class bit_vector {
         m_bits.steal(bits);
     }
 
-    explicit bit_vector(bit_vector_builder* from)
-    {
-        m_size = from->size();
-        m_bits.steal(from->move_bits());
-    }
+    explicit bit_vector(bit_vector_builder* from);
 
     template <typename Visitor>
     void map(Visitor& visit)
