+++ include/pisa/codec/integer_codes.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "bit_vector.hpp"
+#include "bit_vector_builder.hpp"
 
 namespace pisa {
